@@ -44,31 +43,30 @@ license = "MIT"
 packages = [
     { include = "channels_graphql_ws/" },
 ]
-[tool.poetry.dependencies]
+
 # Production dependencies.
 # Version syntax: https://python-poetry.org/docs/dependency-specification/
-python = "^3.6.0"
+[tool.poetry.dependencies]
+python = "^3.8.0"
 aiohttp = "^3"
 asgiref = "^3"
-backports-datetime-fromisoformat = {version = "^1", python = "~3.6"}
-channels = "^3"
-dataclasses = {version = "^0.7", python = "~3.6"}
-Django = "^3"
-graphene = "^2.1"
-graphql-core = ">=2.2"
-msgpack = ">=0.6.1, <2"
-[tool.poetry.dev-dependencies]
+channels = "^4"
+Django = "^4"
+graphene = "^3"
+graphql-core = "3.2.3" # Carefully test after upgrade, since we path private internals.
+msgpack = "^1"
+
 # Development dependencies.
 # Version syntax: https://python-poetry.org/docs/dependency-specification/
-black = { version = "*", allow-prereleases = true }
+[tool.poetry.group.dev.dependencies]
+black = { version = "*" }
+channels = {version = "^4", extras = ["daphne"]}
 channels-redis = "*"
 coverage = "*"
 graphene-django = "*"
-ipython = "*"
 isort = {version = "*", extras = ["pyproject"]}
 mypy = "*"
 plumbum = "*"
-pre-commit = "*"
 pydocstyle = "*"
 pylint = "*"
 pylint-django = "*"
@@ -80,13 +78,12 @@ pytest-django = "*"
 pytest-pythonpath = "*"
 pytest-xdist = "*"
 tox = "*"
-types-dataclasses = {version = "*", python = "~3.6"}
 
 # ---------------------------------------------------------------- BLACK
 # Black - the uncompromising code formatter.
 # https://black.readthedocs.io/en/stable/index.html
 [tool.black]
-target-version = ['py36', 'py37', 'py38']
+target-version = ['py37', 'py38', 'py39', 'py310']
 
 # ---------------------------------------------------------------- ISORT
 # Isort configuration.
