@@ -1,4 +1,4 @@
-# Copyright (C) DATADVANCE, 2010-2021
+# Copyright (C) DATADVANCE, 2010-2023
 #
 # Permission is hereby granted, free of charge, to any person obtaining
 # a copy of this software and associated documentation files (the
@@ -28,8 +28,8 @@ from the Graphene (`graphene/types/mutation.py`).
 import asyncio
 import collections
 import hashlib
-import inspect
 import logging
+from typing import Optional
 
 import asgiref.sync
 import channels.db
@@ -50,9 +50,9 @@ LOG = logging.getLogger(__name__)
 class Subscription(graphene.ObjectType):
     """Subscription type definition.
 
-    Subclass this class to define a GraphQL subscription. The class
-    works with `GraphqlWsConsumer` which maintains a WebSocket
-    connection with the client.
+    Subclass this the Subscription class to define a GraphQL
+    subscription. The class works with the `GraphqlWsConsumer` which
+    maintains a WebSocket connection with the client.
 
     The subclass specifies the following methods. You can define each of
     them as a `@classmethod`, as a `@staticmethod`, or even as a regular
@@ -64,12 +64,18 @@ class Subscription(graphene.ObjectType):
         [async] publish(payload, info, *args, **kwds):
             This method invoked each time subscription "triggers".
             Raising an exception here will lead to sending the
-            notification with the error. To suppress the notification
-            return `Subscription.SKIP`.
+            notification with the error. Technically the WebSocket
+            message will contain extra field "extensions.code" holding
+            the classname of the exception raised. To suppress the
+            notification return `Subscription.SKIP`.
 
-            Can be implemented both as a synchronous or as a coroutine
-            function. In both cases a method runs in a worker thread
-            from the GraphQL-processing threadpool.
+            Can be implemented as both asynchronous (`async def`) or
+            synchronous (`def`) function. Asynchronous implementation
+            runs blazingly fast in the main event loop of the main
+            thread. You must be careful with blocking calls though. You
+            can offload blocking operations to a thread in such cases.
+            Synchronous implementation always runs in a worker thread
+            which comes with a price of extra overhead.
 
             Required.
 
@@ -80,20 +86,24 @@ class Subscription(graphene.ObjectType):
                     information.
                 args, kwds: Values of the GraphQL subscription inputs.
             Returns:
-                The same the any Graphene resolver returns. Returning
-                a special object `Subscription.SKIP` indicates that this
-                notification shall not be sent to the client at all.
+                The same that any Graphene resolver returns. Returning a
+                special object `Subscription.SKIP` indicates that this
+                notification must not be sent to the client at all.
 
         [async] subscribe(root, info, *args, **kwds):
             Called when client subscribes. Define this to do some extra
             work when client subscribes and to group subscriptions into
             different subscription groups. Method signature is the same
-            as in other GraphQL "resolver" methods but it may return
-            the subscription groups names to put the subscription into.
+            as in other GraphQL "resolver" methods but it may return the
+            subscription groups names to put the subscription into.
 
-            Can be implemented both as a synchronous or as a coroutine
-            function. In both cases a method runs in a worker thread
-            from the GraphQL-processing threadpool.
+            Can be implemented as both asynchronous (`async def`) or
+            synchronous (`def`) function. Asynchronous implementation
+            runs blazingly fast in the main event loop of the main
+            thread. You must be careful with blocking calls though. You
+            can offload blocking operations to a thread in such cases.
+            Synchronous implementation always runs in a worker thread
+            which comes with a price of extra overhead.
 
             Optional.
 
@@ -116,9 +126,13 @@ class Subscription(graphene.ObjectType):
             Called when client unsubscribes. Define this to be notified
             when client unsubscribes.
 
-            Can be implemented both as a synchronous or as a coroutine
-            function. In both cases a method runs in a worker thread
-            from the GraphQL-processing threadpool.
+            Can be implemented as both asynchronous (`async def`) or
+            synchronous (`def`) function. Asynchronous implementation
+            runs blazingly fast in the main event loop of the main
+            thread. You must be careful with blocking calls though. You
+            can offload blocking operations to a thread in such cases.
+            Synchronous implementation always runs in a worker thread
+            which comes with a price of extra overhead.
 
             Args:
                 root: Always `None`.
@@ -134,13 +148,11 @@ class Subscription(graphene.ObjectType):
     access to this value as `info.context.zen`.
 
     Static methods of subscription subclass:
-        broadcast: Call this method to notify all subscriptions in the
-            group.
-        unsubscribe: Call this method to stop all subscriptions in the
-            group.
+        broadcast(): Call this to notify all subscriptions in the group.
+        unsubscribe(): Call this to stop all subscriptions in the group.
+
     NOTE: If you call any of these methods from the asynchronous context
     then `await` the result of the call.
-
     """
 
     # ----------------------------------------------------------------------- PUBLIC API
@@ -149,29 +161,27 @@ class Subscription(graphene.ObjectType):
     SKIP = GraphqlWsConsumer.SKIP
 
     # Subscription notifications queue limit. Set this to control the
-    # amount of notifications server keeps in queue when notifications
-    # come faster than server processing them. Set this limit to 1 drops
-    # all notifications in queue except the latest one. Use this only if
-    # you are sure that subscription always returns all required state
-    # to the client and client does not loose information when
-    # intermediate notification is missed.
-    notification_queue_limit = None
+    # amount of notifications server keeps in the queue when
+    # notifications come faster than server processes them. Setting this
+    # to 1 drops all notifications in the queue except the latest one.
+    # Useful to skip intermediate notifications, e.g. progress reports.
+    notification_queue_limit: Optional[int] = None
 
     @classmethod
     def broadcast(cls, *, group=None, payload=None):
         """Call this method to notify all subscriptions in the group.
 
-        This method can be called from both synchronous and asynchronous
-        contexts. If you call it from the asynchronous context then you
-        have to `await`.
+        Can be called from both synchronous and asynchronous contexts.
+
+        It is necessary to `await` if called from the async context.
 
         Args:
             group: Name of the subscription group which members must be
                 notified. `None` means that all the subscriptions of
                 type will be triggered.
             payload: The payload delivered to the `publish` handler.
-                NOTE: The `payload` is serialized before sending
-                to the subscription group.
+                NOTE: The `payload` is serialized before sending to the
+                subscription group.
 
         """
         try:
@@ -180,31 +190,25 @@ class Subscription(graphene.ObjectType):
             pass
         else:
             if event_loop.is_running():
-                assert cls._from_coroutine(), (
-                    "The eventloop is running so this call is going to return"
-                    " a coroutine object, but the function is called from"
-                    " a synchronous context, so you cannot simply 'await' the result!"
-                    " This may indicate a wrong usage. To force some particular"
-                    " behavior directly call 'broadcast_sync' or 'broadcast_async'."
+                return event_loop.create_task(
+                    cls.broadcast_async(group=group, payload=payload)
                 )
-                return cls.broadcast_async(group=group, payload=payload)
 
         return cls.broadcast_sync(group=group, payload=payload)
 
     @classmethod
     async def broadcast_async(cls, *, group=None, payload=None):
         """Broadcast, asynchronous version."""
-        # Offload to the thread cause it do DB operations and may work
-        # slowly.
-        db_sync_to_async = channels.db.database_sync_to_async
-
         # Manually serialize the `payload` to allow transfer of Django
-        # models inside the `payload`.
-        serialized_payload = await db_sync_to_async(Serializer.serialize)(payload)
+        # models inside `payload`, auto serialization does not do this.
+        serialized_payload = await channels.db.database_sync_to_async(
+            Serializer.serialize
+        )(payload)
 
         # Send the message to the Channels group.
         group = cls._group_name(group)
         group_send = cls._channel_layer().group_send
+        # Will result in a call of `GraphqlWsConsumer.broadcast`.
         await group_send(
             group=group,
             message={
@@ -221,10 +225,12 @@ class Subscription(graphene.ObjectType):
         # models inside the `payload`.
         serialized_payload = Serializer.serialize(payload)
 
-        # Send the message to the Channels group.
         group = cls._group_name(group)
-        group_send = asgiref.sync.async_to_sync(cls._channel_layer().group_send)
-        group_send(
+        sync_channel_layer_group_send = asgiref.sync.async_to_sync(
+            cls._channel_layer().group_send
+        )
+        # Will result in a call of `GraphqlWsConsumer.broadcast`.
+        sync_channel_layer_group_send(
             group=group,
             message={
                 "type": "broadcast",
@@ -245,7 +251,6 @@ class Subscription(graphene.ObjectType):
             group: Name of the subscription group which members must be
                 unsubscribed. `None` means that all the client of the
                 subscription will be unsubscribed.
-
         """
         try:
             event_loop = asyncio.get_event_loop()
@@ -253,14 +258,7 @@ class Subscription(graphene.ObjectType):
             pass
         else:
             if event_loop.is_running():
-                assert cls._from_coroutine(), (
-                    "The eventloop is running so this call is going to return"
-                    " a coroutine object, but the function is called from"
-                    " a synchronous context, so you cannot simply 'await' the result!"
-                    " This may indicate a wrong usage. To force some particular"
-                    " behavior directly call 'unsubscribe_sync' or 'unsubscribe_async'."
-                )
-                return cls.unsubscribe_async(group=group)
+                return asyncio.create_task(cls.unsubscribe_async(group=group))
 
         return cls.unsubscribe_sync(group=group)
 
@@ -278,18 +276,32 @@ class Subscription(graphene.ObjectType):
         """Unsubscribe, synchronous version."""
         # Send the message to the Channels group.
         group = cls._group_name(group)
-        group_send = asgiref.sync.async_to_sync(cls._channel_layer().group_send)
-        group_send(group=group, message={"type": "unsubscribe", "group": group})
+        sync_channel_layer_group_send = asgiref.sync.async_to_sync(
+            cls._channel_layer().group_send
+        )
+        sync_channel_layer_group_send(
+            group=group,
+            message={
+                "type": "unsubscribe",
+                "group": group,
+            },
+        )
 
     @classmethod
     def Field(  # pylint: disable=invalid-name
         cls, name=None, description=None, deprecation_reason=None, required=False
     ):
-        """Represent subscription as a field to "deploy" it."""
+        """Represent subscription as a field to mount it to the schema.
+
+        Typical usage:
+            class Subscription(graphene.ObjectType):
+                on_new_chat_message = OnNewChatMessage.Field()
+
+        """
         return graphene.Field(
             cls._meta.output,
             args=cls._meta.arguments,
-            resolver=cls._meta.resolver,
+            resolver=cls._meta.publish,
             name=name,
             description=description,
             deprecation_reason=deprecation_reason,
@@ -308,19 +320,20 @@ class Subscription(graphene.ObjectType):
         arguments=None,
         _meta=None,
         **options,
-    ):  # pylint: disable=arguments-differ
-        """Prepare subscription when on subclass creation.
+    ):  # pylint: disable=arguments-renamed
+        """Prepare subscription on subclass creation.
 
         This method is invoked by the superclass `__init__subclass__`.
         It is needed to process class fields, `Meta` and inheritance
-        parameters. This is genuine Graphene approach.
+        parameters. This is genuine Graphene approach inherited/cloned
+        from the original Mutation class implementation.
         """
         if not _meta:
             _meta = SubscriptionOptions(cls)
 
         output = output or getattr(cls, "Output", None)
         # Collect fields if output class is not explicitly defined.
-        fields = {}
+        fields: dict = {}
         if not output:
             fields = collections.OrderedDict()
             for base in reversed(cls.__mro__):
@@ -333,7 +346,6 @@ class Subscription(graphene.ObjectType):
 
         if not arguments:
             input_class = getattr(cls, "Arguments", None)
-
             if input_class:
                 arguments = graphene.utils.props.props(input_class)
             else:
@@ -342,13 +354,12 @@ class Subscription(graphene.ObjectType):
         # Get `publish`, `subscribe`, and `unsubscribe` handlers.
         subscribe = subscribe or getattr(cls, "subscribe", None)
         publish = publish or getattr(cls, "publish", None)
+        unsubscribed = unsubscribed or getattr(cls, "unsubscribed", None)
         assert publish is not None, (
             f"Subscription '{cls.__qualname__}' does not define a"
             " method 'publish'! All subscriptions must define"
-            " 'publish' which processes a GraphQL query!"
+            " 'publish' which processes GraphQL queries!"
         )
-
-        unsubscribed = unsubscribed or getattr(cls, "unsubscribed", None)
 
         if _meta.fields:
             _meta.fields.update(fields)
@@ -356,92 +367,20 @@ class Subscription(graphene.ObjectType):
             _meta.fields = fields
 
         # Auxiliary alias.
-        get_function = graphene.utils.get_unbound_function.get_unbound_function
+        graphene_get_function = graphene.utils.get_unbound_function.get_unbound_function
 
         # pylint: disable=attribute-defined-outside-init
         _meta.arguments = arguments
         _meta.output = output
-        _meta.resolver = get_function(cls._subscribe)
-        _meta.subscribe = get_function(subscribe)
-        _meta.publish = get_function(publish)
-        _meta.unsubscribed = get_function(unsubscribed)
+        _meta.publish = graphene_get_function(publish)
+        _meta.subscribe = graphene_get_function(subscribe)
+        _meta.unsubscribed = graphene_get_function(unsubscribed)
 
         super().__init_subclass_with_meta__(_meta=_meta, **options)
 
     @classmethod
-    def _subscribe(cls, root, info, *args, **kwds):
-        """Subscription request received.
-
-        This is called by the Graphene when a client subscribes.
-        """
-        # Extract function which associates the callback with the groups
-        # and bring real root back.
-        register_subscription = root.register_subscription
-        root = root.real_root
-
-        # Attach current subscription to the group corresponding to the
-        # concrete class. This allows to trigger all the subscriptions
-        # of the current type, by invoking `publish` without setting
-        # the `group` argument.
-        groups = [cls._group_name()]
-
-        # Invoke the subclass-specified `subscribe` method to get the
-        # groups subscription must be attached to.
-        if cls._meta.subscribe is not None:
-            subclass_groups = cls._meta.subscribe(root, info, *args, **kwds)
-            # Properly handle `async def subscribe`.
-            if asyncio.iscoroutinefunction(cls._meta.subscribe):
-                subclass_groups = asyncio.get_event_loop().run_until_complete(
-                    subclass_groups
-                )
-            assert subclass_groups is None or isinstance(
-                subclass_groups, (list, tuple)
-            ), (
-                f"Method 'subscribe' returned a value of an incorrect type"
-                f" {type(subclass_groups)}! A list, a tuple, or 'None' expected."
-            )
-            subclass_groups = subclass_groups or []
-        else:
-            subclass_groups = []
-
-        groups += [cls._group_name(group) for group in subclass_groups]
-
-        # Register callbacks to call `publish` and `unsubscribed`.
-        # Function `register` provides an observable which must
-        # be returned from here, cause that is what GraphQL expects from
-        # the subscription "resolver" functions.
-        def publish_callback(payload):
-            """Call `publish` with the payload."""
-            result = cls._meta.publish(payload, info, *args, **kwds)
-            # Properly handle `async def publish`.
-            if asyncio.iscoroutinefunction(cls._meta.publish):
-                result = asyncio.get_event_loop().run_until_complete(result)
-            return result
-
-        def unsubscribed_callback():
-            """Call `unsubscribed` notification."""
-            if cls._meta.unsubscribed is None:
-                return None
-            result = cls._meta.unsubscribed(None, info, *args, **kwds)
-            # Properly handle `async def unsubscribed`.
-            if asyncio.iscoroutinefunction(cls._meta.unsubscribed):
-                result = asyncio.get_event_loop().run_until_complete(result)
-            # There is not particular purpose of returning result of the
-            # callback. We do it just for uniformity with `publish` and
-            # `subscribe`.
-            return result
-
-        return register_subscription(
-            groups,
-            publish_callback,
-            unsubscribed_callback,
-            cls.notification_queue_limit,
-        )
-
-    @classmethod
     def _group_name(cls, group=None):
         """Group name based on the name of the subscription class."""
-        prefix = GraphqlWsConsumer.group_name_prefix
         suffix = f"{cls.__module__}.{cls.__qualname__}"
         if group is not None:
             suffix += "-" + group
@@ -451,41 +390,8 @@ class Subscription(graphene.ObjectType):
         # about that the group name is wrong (actually is too long).
         suffix_sha256 = hashlib.sha256()
         suffix_sha256.update(suffix.encode("utf-8"))
-        suffix_sha256 = suffix_sha256.hexdigest()
 
-        return f"{prefix}-{suffix_sha256}"
-
-    @staticmethod
-    def _from_coroutine() -> bool:
-        """Check if called from the coroutine function.
-
-        Determine whether the current function is called from a
-        coroutine function (native coroutine, generator-based coroutine,
-        or asynchronous generator function).
-
-        NOTE: That it's only recommended to use for debugging, not as
-        part of your production code's functionality.
-        """
-        try:
-            frame = inspect.currentframe()
-            if frame is None:
-                return False
-            coroutine_function_flags = (
-                inspect.CO_COROUTINE  # pylint: disable=no-member
-                | inspect.CO_ASYNC_GENERATOR  # pylint: disable=no-member
-                | inspect.CO_ITERABLE_COROUTINE  # pylint: disable=no-member
-            )
-            if (
-                frame is not None
-                and frame.f_back is not None
-                and frame.f_back.f_back is not None
-            ):
-                return bool(
-                    frame.f_back.f_back.f_code.co_flags & coroutine_function_flags
-                )
-            return False
-        finally:
-            del frame
+        return f"{GraphqlWsConsumer.group_name_prefix}-{suffix_sha256.hexdigest()}"
 
     @classmethod
     def _channel_layer(cls):
@@ -504,3 +410,4 @@ class SubscriptionOptions(graphene.types.objecttype.ObjectTypeOptions):
     output = None
     subscribe = None
     publish = None
+    unsubscribed = None
