@@ -1,4 +1,4 @@
-# Copyright (C) DATADVANCE, 2010-2021
+# Copyright (C) DATADVANCE, 2010-2023
 #
 # Permission is hereby granted, free of charge, to any person obtaining
 # a copy of this software and associated documentation files (the
@@ -91,7 +91,7 @@ INSTALLED_APPS: List[str] = [
     "django.contrib.admin",
     "channels",
 ]
-ALLOWED_HOSTS = "*"
+ALLOWED_HOSTS = ["*"]
 STATIC_URL = "/static/"
 STATICFILES_FINDERS = ["django.contrib.staticfiles.finders.AppDirectoriesFinder"]
 # In this simple example we use in-process in-memory Channel layer.
