@@ -1,4 +1,4 @@
-# Copyright (C) DATADVANCE, 2010-2021
+# Copyright (C) DATADVANCE, 2010-2023
 #
 # Permission is hereby granted, free of charge, to any person obtaining
 # a copy of this software and associated documentation files (the
@@ -100,13 +100,15 @@ async def test_main_usecase(gql):
     # Mutation response.
     resp = await client.receive(assert_id=msg_id, assert_type="data")
     assert resp["data"] == {"send_chat_message": {"message": message}}
-    await client.receive(assert_id=msg_id, assert_type="complete")
+    resp = await client.receive(assert_id=msg_id, assert_type="complete")
 
+    print("Check that subscription notification were sent.")
     # Subscription notification.
     resp = await client.receive(assert_id=sub_id, assert_type="data")
     event = resp["data"]["on_chat_message_sent"]["event"]
     assert json.loads(event) == {
-        "user_id": UserId.ALICE,
+        # pylint: disable=no-member
+        "user_id": UserId.ALICE.value,  # type: ignore[attr-defined]
         "payload": message,
     }, "Subscription notification contains wrong data!"
 
@@ -197,7 +199,7 @@ async def test_subscribe_unsubscribe(gql):
                 }
                 """
             ),
-            "variables": "",
+            "variables": {},
             "operationName": "op_name",
         },
     )
@@ -330,7 +332,7 @@ async def test_subscription_groups(gql):
     await trigger_subscription(comm_alice, tom_id, message)
     # Check Tom's notifications.
     resp = await comm_tom.receive(assert_id=uid_tom, assert_type="data")
-    check_resp(resp, UserId[tom_id].value, message)
+    check_resp(resp, UserId[tom_id].value, message)  # type: ignore[misc]
     # Any other did not receive any notifications.
     await comm_alice.assert_no_messages()
     print("Trigger subscription: send message to Alice.")
@@ -338,7 +340,7 @@ async def test_subscription_groups(gql):
     await trigger_subscription(comm_tom, alice_id, message)
     # Check Alice's notifications.
     resp = await comm_alice.receive(assert_id=uid_alice, assert_type="data")
-    check_resp(resp, UserId[alice_id].value, message)
+    check_resp(resp, UserId[alice_id].value, message)  # type: ignore[misc]
     # Any other did not receive any notifications.
     await comm_tom.assert_no_messages()
 
@@ -348,9 +350,9 @@ async def test_subscription_groups(gql):
 
     print("Check Tom's and Alice's notifications.")
     resp = await comm_tom.receive(assert_id=uid_tom, assert_type="data")
-    check_resp(resp, UserId[tom_id].value, message)
+    check_resp(resp, UserId[tom_id].value, message)  # type: ignore[misc]
     resp = await comm_alice.receive(assert_id=uid_alice, assert_type="data")
-    check_resp(resp, UserId[alice_id].value, message)
+    check_resp(resp, UserId[alice_id].value, message)  # type: ignore[misc]
 
     print("Disconnect and wait the application to finish gracefully.")
     await comm_tom.finalize()
@@ -424,7 +426,7 @@ class OnChatMessageSent(channels_graphql_ws.Subscription):
     def publish(self, info, user_id):
         """Publish query result to the subscribers."""
         del info
-        event = {"user_id": user_id, "payload": self}
+        event = {"user_id": user_id.value, "payload": self}
 
         return OnChatMessageSent(event=event)
 
